more thoughts
